Ignore status code 503
Spring actuator returns status code 503 on the health endpoint when it reports status DOWN. This should be seperated from other possible http exceptions.
It can be argued that the health endpoint should return the status code 200 since the prupose of the endpoint is to report the health in the json payload and not report the health using HTTP status code.